pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "encycloscope-mcp"
version = "1.0.0"
description = "MCP Server for exploring the Encyclopédie de Diderot et d'Alembert"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Morgan Blangeois" }
]
keywords = ["mcp", "encyclopedie", "diderot", "dalembert", "philosophy", "18th-century"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mcp>=1.0.0",
    "fastmcp>=0.1.0",
    "uvicorn>=0.30.0",
    "starlette>=0.38.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Semantic search with d'AlemBERT + FAISS
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "faiss-cpu>=1.7.0",
    # Legacy support
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
gpu = [
    "faiss-gpu>=1.7.0",  # For GPU-accelerated indexing
]

[project.scripts]
encycloscope-mcp = "encycloscope_mcp.http_streamable:main"

[tool.hatch.build.targets.wheel]
packages = ["src/encycloscope_mcp"]
